对用湖账户，充值记录、账单 测试，编写测试用例

+++ cloudplatform-webportal-sdk-billing/src/main/java/com/letv/portal/dao/letvcloud/BillRechargeRecordMapper.java
@@ -9,8 +9,8 @@ import com.letv.portal.model.letvcloud.BillRechargeRecord;
  * Created by wanglei14 on 2015/6/29.
  */
 public interface BillRechargeRecordMapper {
-    Long insert(BillRechargeRecord record);
-    Long updateSuc(Map<String,Object> params);
+    int insert(BillRechargeRecord record);
+    int updateSuc(Map<String,Object> params);
     BillRechargeRecord getAmount(String tradeNum);
     List<BillRechargeRecord> getUserAmountRecord(Map<String,Object> params);
     int getAddRecordCnt(Long userId);
